MT4Connector/requirements.txt
requests
python-dateutil
pytz
watchdog
python-telegram-bot==13.7